[#25114] DocDB: Handle dropped table when applying intents
Summary:
When colocated table is dropped, the table continue to run.
As result we could try to apply transaction on this tablet, that was committed before table was dropped.
But we expect that this table always exists if packed row is used.
Adjusted logic to handle dropped table correctly.

Also renamed ExternalIntentsBatchWriter to  NonTransactionalBatchWriter.
Since ExternalIntentsBatchWriter name is incorrect.
External intents handling just small part of this writer, moreover this part usually is not used.
Jira: DB-14261

Test Plan: PgPackedRowTest.DropColocatedTable/*

Reviewers: hsunder, slingam, rthallam

Reviewed By: hsunder, slingam, rthallam

Subscribers: rthallam, ybase, yql

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D40598