[PLAT-10615]Parallelize support bundle creation
Summary:
Currently all components of support bundle are collected sequentially in a single thread.
Added changes to add all components to a subtask group to be ran in parallel.

This diff is part of the changes planned for the next phase of support bundle.
More details in the [[ https://docs.google.com/document/d/1B1gkJiTVkxN5OAKVlvCIbEAWTYw6SCCtaFok-zpD6eM/edit?usp=sharing | Design doc ]].

Test Plan: Manually verified that collection of all components happens in parallel and the contents of the Support bundle are as expected.

Reviewers: #yba-api-review!, nsingh, skurapati, anijhawan

Reviewed By: nsingh, anijhawan

Subscribers: anijhawan, yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D40620