[BACKPORT 2024.1][#25250] Fix index inconsistency caused by concurrent writes during index backfill
Summary:
When creating an index, the online schema change process reaches the
backfilling phase and picks a backfill read time. During this backfill
stage, main table entires are read and index entries are inserted with a
write time same as the backfill read time. It is important for the write
time of backfilled entries to be <= the backfill read time so that the
backfilled entries don't mask newer writes which are not part of the
backfill read snapshot.

However, due to #25250, many index entries are being backfilled with write
times using the current hybrid time on docdb instead of the backfill read
time. This can result in inconsistency bugs like the following:

  -> Consider a table (k int primary key, v int)
  -> Index creation on (v) is issued
  -> Backfill read time is picked
  -> Some connection DELETEs row with v=9999 on main table (write_time_a)
  -> Row with index key 9999 is read (as of the backfill snapshot)
  -> Row with index key 9999 is backfilled (write_time_b)

If write_time_b > write_time_a, the index will have row with index key 9999
but the main table will not.

# Cause:
This happens because a pg backend performing backfill sends
a single Perform rpc with multiple ops to the same node's PgClientService
along with the `write_time` field set (see PgPerformRequestPB). This service
calls PrepareOperations() in pg_client_session.cc which creates a YBPgsqlWriteOp
for each write operation from PG. These ops are multiplexed
by the batcher into separate rpcs based on the remote tservers they are
destined for. Each remote tserver rpc is prepared in the
async_rpc.cc file, which calls HandleExtraFields() that relies on the
`write_time_` of the YBPgsqlWriteOps to set the external_hybrid_time field for
the write rpc.

For an rpc to have the correct write time, at least one YBPgsqlWriteOp
in it should have the write_time set. However, PrepareOperations() in pg_client_session.cc
only sets the `write_time_` for the first op, leading to only one remote tserver
getting a write_time.

So in effect, if a backfill operation on a PG backend reads main table rows
that are to be backfilled into N index tablets, only one of these tablets will
have writes with the correct write_time (i.e., same as the backfill read
time).

This buggy behaviour is a regression caused by commit
c5f512583fc7ecbc054584d59b19c584a94bd0ee which was first
part of release 2.13.1.0. Before this commit, the write times of
backfilled index entries were the same as the backfill read time
across all index tablets.

# Fix:
Modify PrepareOperations() to set the `write_time_` in all YBPgsqlWriteOps
created for ops in the Perform rpc.

Original commit: ea73ed26f9820d950c22d1e76da00c1280ee8108 / D40733

Test Plan: Jenkins

Reviewers: jason, mihnea

Reviewed By: mihnea

Subscribers: mihnea, yql, ybase

Differential Revision: https://phorge.dev.yugabyte.com/D40842