[#25276] docdb: Vector index creation fails for a column with id different from 1
Summary:
Vector index creation leads to a crash in debug mode if vector column id is different from 1. And
it logs an error in release mode. This happens because index table schema gets a wrong column id
during marking the column as vector column -- the value comes from the indexable table's
vector column and there's no way to get that information for the index table.

The fix is to mark the last column of index table as a vector column always, as it is expected
that vector index schema always consists of 2 columns only: ybidxbasectid and <vector column>.

This logic will be removed once new DataType::VECTOR is implemented.

The bug was brought with https://phorge.dev.yugabyte.com/D40397.
Jira: DB-14462

Test Plan: Jenkins

Reviewers: slingam, sergei

Reviewed By: slingam

Subscribers: ybase

Differential Revision: https://phorge.dev.yugabyte.com/D40651