[PLAT-16274] Restart the node_exporter service
Summary:
We recently moved the metric collection directory under `/home/yugabyte` which prometheus user was not able to read.
As part of this diff - https://github.com/yugabyte/yugabyte-db/commit/9fd7781fe954cc5b5cc76c850babdad6d8a1efad we added prometheus user to `yugabyte` group, which was not getting reflected.
We needed the restart of `node_exporter` service - This diff handles the same.

Test Plan:
Created a universe with node_exporter running as prometheus user.
Ran the above playbook, verified that the systemd unit is modified to yugabyte

Reviewers: anijhawan, muthu, daniel, yshchetinin

Reviewed By: daniel

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D40672