[PLAT-16326] Kubernetes Pause Resume does not need overrides file
Summary:
Because resume is using --reuse-values, it will reuse all the values of
the previous helm release (which, is from before the pause). Pause does
not create a new helm release, and instead edits the stateful set
directly using kubectl.

In fact, the overrides file was not getting generated correctly either,
causing ybc.enabled to always be false and masterAddresses to get passed
as null.

Test Plan:
tested multiple pause resumes
1. master addresses always set
2. ybc is enabled when its suppose to be

Reviewers: anijhawan, #yba-api-review

Reviewed By: anijhawan, #yba-api-review

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D40682