[PLAT-13317] Do not allow skipping stopping processes unless we are pausing the universe
Summary:
For many tasks, we do not want to skip stop control commands if the instance state is not running. This can be an issue as if the VM is started/not paused, the master/tserver/yb-controller processes will automatically start up again. This diff fixes this issue to not skip the 'stop' control command (may cause task to fail, but is expected) unless it is a `pauseUniverse` as YBM needs this to retry pause universe and the VM may be paused if it failed on previous runs.

Also, combined some of the methods to reduce amount of code duplication

Test Plan:
Create a 3 node rf3 universe.

Simulate 2 scenarios
1. First pause one of the nodes on the universe. Then perform a rolling restart. Make sure that the restart fails.
2. Perform a pause universe with a fault injection after stopping the VM. Re-run a new pause universe task, making sure that the processes have skipped the stop command.

Reviewers: nsingh, vbansal

Reviewed By: nsingh

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D40684