Merged PR 1524148: [Bugfix] Fix handling of compound wildcard text indexes
### Does this PR have any customer impact?
Yes

### Type (Feature, Refactoring, Bugfix, DevOps, Testing, Perf, etc)
Bugfix

### Does it involve schema level changes? (Table, Column, Index, UDF, etc level changes)
No

### Are you introducing any new config? If yes, do you have tests with and without them being set?
No

### Release Note (Refer [Template](../docs/release_notes/pgmongo_template.md))

### Description
Handling of wildcard text indexes when compounded with a non-wildcard field is broken. This bug is a **functional bug** affecting query results when used with a wildcard index (see JS test for repro).

----
#### AI description  (iteration 1)
#### PR Classification
Bug fix to correct the handling of compound wildcard text indexes.

#### PR Summary
This pull request addresses a bug in the handling of compound wildcard text indexes by adding additional tests and correcting the index creation and listing logic.
- Modified `commands_create_indexes_text.out` to include more tests and correct the handling of compound wildcard text indexes.
- Added new test file `text_indexes_wildcard.js` to verify the correct behavior of wildcard text indexes.
- Updated `commands_create_indexes_text.sql` to include additional test cases for compound wildcard text indexes.