[#25268] YSQL: Simplify ExecLockRows code
Summary:
The `ExecLockRows` function has the code to check that row locks are not requested for YB and non-YB tables at same time. It is reasonable to move this code into initialization step of the  Lock Rows node to avoid wasting of CPU cycles on each row processing.
Jira: DB-14457

Test Plan: Jenkins

Reviewers: pjain

Reviewed By: pjain

Subscribers: yql

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D40632