Make default numeric scale 9 (#86)
To give the integral part more room, we decided to make the default scale = 9
when the user does not specify precision and scale for numeric type.

Multiplication of 2 numeric doubles the result scale and gives the integral part
less digits. It is wiser to make it smaller but still useful number.