[#25335] docdb: Vector index: DFATAL crash during copying of large rows
Summary:
The issue happens when inserted row is larger than a configured packed row limit (32_KB). In this
case the remaining part of the row is being inserted as a non-packed row. When such a row is being
processed, the vector index inserter does not check that a vector’s value can be null for a
packed row and tries to insert it into the vector index, which lead to the insertion failure at
vector index side and subsequent DFATAL.

This change adjusts the logic of packed rows insertion into a vector index: before the insertion
into a vector index, it now checks if vector's value is not a null value.
Jira: DB-14548

Test Plan: Jenkins

Reviewers: slingam, sergei

Reviewed By: slingam, sergei

Subscribers: ybase

Differential Revision: https://phorge.dev.yugabyte.com/D40750