[PLAT-16259] xCluster upgrade tests are failing tests are failing at Resume after upgrade
Summary:
The diff updates the `lockUniverseIfExist` method to set checkSuccess to true. This is required because during software upgrade of the target universe of an xCluster config, the source universe could be locked to transfer the required tls certificates, but if checkSuccess is not true, the placementModificationTaskUuid won't be reset to null.

When placementModificationTaskUuid is not null, it means there was a previously tasked failed on the universe and until that succeeds, the user cannot run other tasks on the universe.

Test Plan: Rerunning the failing itest with this change and ensure it passes.

Reviewers: nsingh, cwang, nbhatia

Reviewed By: nsingh

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D40777