[#24834] YSQL: Enable INSERT ON CONFLICT read batching with NULLS NOT DISTINCT unique indexes
Summary:
D38354 and D39023 introduced read-batching support for INSERT ON CONFLICT queries. D39058 introduced NULLS NOT DISTINCT support for unique indexes but disabled INSERT ON CONFLICT read-batching if the relation has such an index. This revision enables that.

INSERT ON CONFLICT read-batching uses the QL_OP_IN operator to look up conflicts in a batched fashion.

Before this revision, yb_scan filtered out NULL values (or rows containing NULL columns) from the RHS of QL_OP_IN, and there was no mechanism to skip this filtering. This was so because there was no use case to scan for nulls in this fashion. However, this is required to check for conflicting null values in a nulls-not-distinct unique index in the above-described batched fashion. This revision adds it.

This revision is co-authored by @kramanathan.
Jira: DB-13945

Test Plan:
  ./yb_build.sh --java-test 'org.yb.pgsql.TestPgRegressInsertOnConflict'
  ./yb_build.sh --java-test 'org.yb.pgsql.TestPgRegressInsertOnConflictBatch1'
  ./yb_build.sh --java-test 'org.yb.pgsql.TestPgRegressInsertOnConflictBatch2'
  ./yb_build.sh --java-test 'org.yb.pgsql.TestPgRegressInsertOnConflictBatch3'
  ./yb_build.sh --java-test 'org.yb.pgsql.TestPgRegressInsertOnConflictBatch9'

Reviewers: jason, kramanathan, tnayak

Reviewed By: jason

Subscribers: kramanathan, yugaware, aagrawal, yql, ybase, smishra

Differential Revision: https://phorge.dev.yugabyte.com/D39852