[BACKPORT 2024.2][#25285] DocDB, ASH: Pass and adopt the wait state ptr in TransactionReady
Summary:
Original commit: 04e0328e0b78e268fbbaea3bde7f5968bd88398c / D40597
The wait state pointer is not passed and adopted when a
transaction is processed from the waiters list. So, when the
scope of ADOPT_WAIT_STATE is finished, the metadata is
deleted.

This diff fixes this by passing the wait state pointer and
adopting the wait state pointer.
Jira: DB-14485

Test Plan:
./yb_build.sh --cxx-test geo_transactions_promotion-test --gtest_filter AshGeoPartitionedTransactionTest.TestNonEmptyAshMetadata -n 100

The repro mentioned in the GHI cannot be repro-ed from the unit test, so instead we used geo promoted txns which goes through the same code path

Reviewers: amitanand, hbhanawat, esheng

Reviewed By: amitanand

Subscribers: smishra, ybase, yql

Differential Revision: https://phorge.dev.yugabyte.com/D40794