[PLAT-14899] Add UI support for pushing node agent to legacy nodes
Summary:
This diff adds support for pushing node agent to legacy nodes.
On the universe action menu, it adds an "more" submenu which has a new "Install Node Agent" action.
 - The menu item displays "Install Node Agent" when the universe has node agent installation pending and "Reinstall Node Agent" if it already has attempted node agent installation.
    This is determined by the `installNodeAgent` field in the universeDetails returned from the backend. If `installNodeAgent` = true, then it means we still need to install node agent.
We also moved the following universe actions under the "more" submenu:
  - edit ysql/ycql configuration
  - edit postgres compatibility
  - edit connection pooling
  - disable backup
  - edit security
This is done to reduce the vertical height of the universe actions menu through grouping the less commonly used actions under
a submenu.

On the universe dashboard, it adds a collapsible but non-dismissable banner if any node agent for a running universe has errors and
the global runtime config `yb.node_agent.enabler.scan_interval` being greater than 0 minutes.
This diff adds a localStorage key `isInstallNodeAgentReminderBannerExpanded` which persists the expand/collapse state of the banner.
This is done so users don't need to close the banner every time they visit that page if they've already acknowledged the content.

On the node agent page, for each node agent row also we add a install node agent action.
Both the universe action menu item and the per node action item from the /nodeagent page open the same
'Install Node Agent' modal.
In both cases, the user is present with a choice to either install node agent for all nodes in the same universe,
or a specific node. If the modal was opened from the node agent page, then the specific node is restricted to the
node from which the modal was opened (see screenshot in test plan).

Test Plan:
Verify the user is able to (re)install node agent for a universe from the universe action menu.
{F319049}
{F319112}
{F319114}

Verify the user is able to (re)install node agent for a particular node from the node agent page.
{F319115}
Verify the universe dashboard warning appears as expected when the conditions are met.
{F319319}
{F319320}

Verify the UI toggle for `yb.node_agent.client_enabled` runtime config also comes with a warning message
informing the user about future YBA upgrade requirements.
{F319426}
{F319427}

Reviewers: rmadhavan, nsingh, mkazerooni

Reviewed By: nsingh

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D39851