[#15885] Fix flaky test PgMiniTest.TabletSplitSecondaryIndexYSQL
Summary: The test expects the main table to split during the test. This doesn't always happen because the sst files generated are less than the force split threshold set in the test. Fixing it to 10K should cause it split more predictably. This gflag is only used for this one test, so it shouldn't affect other tests.

Test Plan:
Run the test a few times.

```
./yb_build.sh release --cxx-test pg_mini-test --gtest_filter PgMiniTest.TabletSplitSecondaryIndexYSQL
```

Reviewers: asrivastava

Reviewed By: asrivastava

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D40801