[#25393] xCluster: Fix XClusterYsqlTest.DropTableOnProducerOnly in ASAN
Summary:
The sleep timeout is calculated on global static initialization time which gets the default value of `cdc_parent_tablet_deletion_task_retry_secs` which is 30. The test cases set this flag to a lower number but do so much later during test execution. This change replaces the global static with a function that will compute the duration on the fly.

Also fixing ASAN issues with `XClusterYSqlTestConsistentTransactionsTest` due to low timeouts

Fixes #25393
Jira: DB-14622

Test Plan:
XClusterYsqlTest.DropTableOnProducerOnly
XClusterYSqlTestConsistentTransactionsTest

Reviewers: jhe, sergei, xCluster

Reviewed By: jhe

Subscribers: ybase

Differential Revision: https://phorge.dev.yugabyte.com/D40836