[#25336] YSQL: Disable check_is_install_user
Summary:
Vanilla PG expects one user during the upgrade, the bootstrap super user. This defaults to
`postgres`, but can be set with the `-U` argument to `initdb`.

In Yugabyte, the upgrade might be run by different Yugabyte users. Rather than try to capture every possible yugabyte user and rewriting the check to that, we can just ignore the user check. Since we control the creation of the new cluster, we don't have to be as vigilant against weird cases as Postgres does.

Fixes #25336
Jira: DB-14549

Test Plan:
```
./yb_build.sh release --cxx-test pg15_upgrade-test
```

Reviewers: hsunder

Reviewed By: hsunder

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D40726