[BACKPORT 2024.2] Revert "[BACKPORT 2024.2][#24106][#24107] CDC: Support lsn_type in CREATE_REPLICATION_SLOT ddl"
Summary:
This diff reverts the commit a672c02fcba1386c780da54c4be353042cf305a7 on 2024.2, which added a field to `PgCreateReplicationSlotRequestPB` with a different field number than in master. This would cause upgrade issues.

**Upgrade/Rollback safety:**
The revert is not technically upgrade-safe, but is required for long-term upgradability. If there are any releases with this diff NOT reverted, then upgrading from that release to this one might causes undefined behavior including crashes.
Jira: DB-12995, DB-12996

Test Plan: jenkins:urgent

Reviewers: hsunder, asaha, asrivastava, xCluster

Reviewed By: hsunder, asrivastava

Subscribers: ycdcxcluster

Differential Revision: https://phorge.dev.yugabyte.com/D40892