[PLAT-16389]Certificate verification fails due to formatting differences between OpenSSL versions
Summary:
Earlier check for (RSA Public-Key) doesn't work for newer openssl version (3.0 & above). Instead of relying on Public-Key, we are now greping the Signature Algorithm directly which is present and consistent across all openssl versions. The validation passes if the signature algorithm has rsa in it.
Added an additional check to skip comparing the md5 checksums if md5sum util is not present in the node.

Test Plan:
Tested locally by creating a universe on ubuntu 22 (openssl version is 3.0)

Without this change create universe fails with

```
 ybops.common.exceptions.YBOpsRuntimeError: Runtime error: Certificate verification for RootCA certificate on node - yb-admin-anab-onprem-create-n1 failed with error: Runtime error: Root certificate - /opt/yugabyte/keys/cert/ca.crt is not using RSA algorithm.
```
With this change, the validation passes.
Verified all validations on openssl v1.0.2, v3.0.2 and v3.2.2
Tried universe creation on Almalinux8 and AmazonLinux2 as well for any regressions

Reviewers: asharma

Reviewed By: asharma

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D40904