[#8096] YSQL: Fix flaky test TestPgDdlConcurrency#testModifiedTableWrite
Summary:
The test is runs a DDL thread concurrently with 10 DML threads in ASAN/TSAN
builds and 50 DML threads in other builds. With 50 threads it still takes too
much resources, with 10 threads it is more flaky.

This diff makes two changes:
(1) Use 10 DML threads for all builds to reduce test resources.
(2) Changed the test to run until one of the expected errors appear to make it
more reliable.
Jira: DB-4809

Test Plan:
On AlmaLinux:
./yb_build.sh release --java-test 'org.yb.pgsql.TestPgDdlConcurrency#testModifiedTableWrite' -n 100 --tp 1
./yb_build.sh asan --java-test 'org.yb.pgsql.TestPgDdlConcurrency#testModifiedTableWrite' -n 100 --tp 1
./yb_build.sh tsan --java-test 'org.yb.pgsql.TestPgDdlConcurrency#testModifiedTableWrite' -n 100 --tp 1
./yb_build.sh debug --java-test 'org.yb.pgsql.TestPgDdlConcurrency#testModifiedTableWrite' -n 100 --tp 1
./yb_build.sh fastdebug --gcc11 --java-test 'org.yb.pgsql.TestPgDdlConcurrency#testModifiedTableWrite' -n 100 --tp 1

On MacOS:
./yb_build.sh release --java-test 'org.yb.pgsql.TestPgDdlConcurrency#testModifiedTableWrite' -n 100 --tp 1

Reviewers: hsunder

Reviewed By: hsunder

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D40929